requirements.txt
# Core RAG and LLM libraries
langchain>=0.1.0
chromadb>=0.4.0
openai>=1.0.0
tiktoken>=0.5.0

# NLP and Text Processing
nltk>=3.8
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# PDF Processing
pypdf>=3.0.0

# Web Scraping (for Wikipedia fetching)
beautifulsoup4>=4.12.0
requests>=2.31.0

# Data Processing and Visualization
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
